gcp/requirements.txt
tensorflow==2.8.0
google-cloud-storage==1.16.1
Pillow==6.0.0
Werkzeug==2.0
tensorflow-hub>=0.6.0